+++ lib/project_types/node/messages/messages.rb
@@ -21,9 +21,9 @@ module Node
             Usage: {{command:%s create node}}
             Options:
               {{command:--name=NAME}} App name. Any string.
-              {{command:--app_url=APPURL}} App URL. Must be a valid URL.
-              {{command:--organization_id=ID}} Partner organization ID. Must be an existing organization.
-              {{command:--shop_domain=MYSHOPIFYDOMAIN }} Development store URL. Must be an existing development store.
+              {{command:--app-url=APPURL}} App URL. Must be a valid URL.
+              {{command:--organization-id=ID}} Partner organization ID. Must be an existing organization.
+              {{command:--shop-domain=MYSHOPIFYDOMAIN }} Development store URL. Must be an existing development store.
           HELP
           error: {
             node_required: "node is required to create an app project. Download at https://nodejs.org/en/download.",