+++ lib/project_types/rails/messages/messages.rb
@@ -30,11 +30,11 @@ module Rails
             Usage: {{command:%s create rails}}
             Options:
               {{command:--name=NAME}} App name. Any string.
-              {{command:--app_url=APPURL}} App URL. Must be a valid URL.
-              {{command:--organization_id=ID}} Partner organization ID. Must be an existing organization.
-              {{command:--shop_domain=MYSHOPIFYDOMAIN }} Development store URL. Must be an existing development store.
+              {{command:--app-url=APPURL}} App URL. Must be a valid URL.
+              {{command:--organization-id=ID}} Partner organization ID. Must be an existing organization.
+              {{command:--shop-domain=MYSHOPIFYDOMAIN }} Development store URL. Must be an existing development store.
               {{command:--db=DB}} Database type. Must be one of: mysql, postgresql, sqlite3, oracle, frontbase, ibm_db, sqlserver, jdbcmysql, jdbcsqlite3, jdbcpostgresql, jdbc.
-              {{command:--rails_opts=RAILSOPTS}} Additional options. Must be string containing one or more valid Rails options, separated by spaces.
+              {{command:--rails-opts=RAILSOPTS}} Additional options. Must be string containing one or more valid Rails options, separated by spaces.
           HELP
 
           error: {