+++ test/project_types/node/commands/create_test.rb
@@ -215,12 +215,20 @@ module Node
 
       private
 
-      def perform_command
+      def perform_command_snake_case
         run_cmd("create node \
           --name=test-app \
           --type=public \
           --organization_id=42 \
           --shop_domain=testshop.myshopify.com")
+      end
+
+      def perform_command
+        run_cmd("create node \
+          --name=test-app \
+          --type=public \
+          --organization-id=42 \
+          --shop-domain=testshop.myshopify.com")
       end
 
       def expect_node_npm_check_commands